Separate get variables logic into a utils file
